Update clusters_evaluation.R
Trying to use MCSC for bacterial genome assembly decontamination (I hear you, this tool was not design for that), I was getting the following error at the line "rec_cluster_tree(tab,which(tab$Cluster=="1"), index = "WR")":

 Error in if (!limite || data$group[cl] == limite) { : 
  missing value where TRUE/FALSE needed 

Adding "!limite==TRUE" in the if prevents an error from the R script, and makes sure that the rest is executed (i.e. the fasta files are created). This allow the MCSC_decontamination.sh to complete.

Although this solves my problem, I can't really tell if it would have an impact for other types of input file. To check!